Commented out the wxPanel and wxGuage. Made a small change to updater.cpp to keep from getting a warning about an undeclared variable.
Please note that the DownloadFile.cpp and DownloadFile.h
do trigger and starts the download. I added the wxPanel and wxGuage in main to get an idea of what they would look like and they should trigger in BrowserFrame::OnFileImageDownload
NOT as in the place commented.

The Guage should pull from nFileSize and nDownloaded during OnFileImageDownloadEvent