Updater was broken and not calling the white-wolf website like it should.  Repaired. I tried to keep the same format.
FileDownload in main had a small bug in the line that allowed downloads for 1 file.  It was calling from the database quite right.  Repaired.

Also now after I fixed these and attempted to view an image it didnt crash anymore.  It now just brings up a invalid zip file like it should.