This is the inventory fix that I have been working on.
It adds a column for Set when the inventory button is pushed.
And also sorts the cards by all the sets.

This also stops the bug in the inventory code mentioned in the 
Defect 68.  Because it didnt sort by set it caused an error to the writing of the database.  So now that it sorts by set it fixes the problem.