Fix the incorrect logic, which assumes that the icon files folder is inside the database directory.
Icon files are always inside the ARDB package folder. On the other hand, the database file is inside the user's home directory. When Database::Instance() is called the first time in MyApp::OnInit() it changes the currently working directory of the program to where the database file is located. Icon files are accessed using relative addressing, which, because CWD is changed, looks under the database folder and fails.

As a solution, save the CWD at the beginning of the program to a global variable and use absolute paths when accessing the icon files.